Re-organise loading benchmarks.
